hashlib throws strange errors when running legacy Python code
The logic behind this code is for this Python version to keep running on legacy systems. The hashlib is likely to break when maintaining legacy Python code and this will likely never be fixed. Therefore ignore errors from c libs on import and send exception when actually trying to use the hashlib methods instead of when importing the C library.